+++ client/js/modals/admin-users-create.js
@@ -1,4 +1,4 @@
-/* global $, Vue */
+/* global $, Vue, alerts */
 
 // Vue Create User instance
 
@@ -8,7 +8,8 @@ let vueCreateUser = new Vue({
     email: '',
     provider: 'local',
     password: '',
-    name: ''
+    name: '',
+    loading: false
   },
   methods: {
     open: (ev) => {
@@ -21,7 +22,28 @@ let vueCreateUser = new Vue({
       vueCreateUser.provider = 'local'
     },
     create: (ev) => {
-      vueCreateUser.cancel()
+      vueCreateUser.loading = true
+      $.ajax('/admin/users/create', {
+        data: {
+          email: vueCreateUser.email,
+          provider: vueCreateUser.provider,
+          password: vueCreateUser.password,
+          name: vueCreateUser.name
+        },
+        dataType: 'json',
+        method: 'POST'
+      }).then((rData, rStatus, rXHR) => {
+        vueCreateUser.loading = false
+        if (rData.ok) {
+          vueCreateUser.cancel()
+          window.location.reload(true)
+        } else {
+          alerts.pushError('Something went wrong', rData.msg)
+        }
+      }, (rXHR, rStatus, err) => {
+        vueCreateUser.loading = false
+        alerts.pushError('Error', rXHR.responseJSON.msg)
+      })
     }
   }
 })